+++ src/test/java/org/la4j/matrix/AbstractMatrixTest.java
@@ -17,6 +17,7 @@
  * 
  * Contributor(s): Evgenia Krivova
  *                 Jakob Moellers
+ *                 Maxim Samoylov
  * 
  */